pytorch: testing new model to scale on azure (#1)
* pytorch: testing new model to scale on azure

problem: the current resnet model is not scaling/running
at a reasonable rate. This start of work will run 8 epochs
on 1 or 2 nodes at reasonable times. We should work on it
together to do final tweaks for the experiment. See the
todo items in the README under the docker/pytorch directory.
The container has been built and pushed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* finishing up testing

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>